lesson3 done2

+++ javalessons/src/main/java/turlo/lesson3/Employee.java
@@ -4,11 +4,11 @@ public class Employee {
     private String name;
     private String post;
     private String email;
-    private long telephone;
-    private long salary;
+    private int telephone;
+    private int salary;
     private int age;
 
-    public Employee(String name, String post, String email, long telephone, long salary, int age) {
+    public Employee(String name, String post, String email, int telephone, int salary, int age) {
         this.name = name;
         this.post = post;
         this.email = email;
